move mempool around
